#79 - Moved the PDF/UA test cases to JUnit tests.
At least confirm that the PDF/UA implementation doesn't throw unexpected exceptions.
